+++ database/migrations/2019_04_03_044839_create_impactos_table.php
@@ -15,7 +15,7 @@ class CreateImpactosTable extends Migration
     {
         Schema::create('impactos', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->float('cantidad');
+            $table->double('cantidad');
             $table->bigInteger('evento_id')->unsigned();
             $table->bigInteger('tipo_impacto_id')->unsigned();
             $table->foreign('evento_id')->references('id')->on('eventos')->onDelete('restrict');
